+++ griptape/common/reference.py
@@ -5,7 +5,7 @@ from typing import Optional
 
 from attrs import Factory, define, field
 
-from griptape.mixins import SerializableMixin
+from griptape.mixins.serializable_mixin import SerializableMixin
 
 
 @define(kw_only=True)