+++ tests/mocks/mock_futures_executor.py
@@ -1,4 +1,4 @@
-from griptape.mixins import FuturesExecutorMixin
+from griptape.mixins.futures_executor_mixin import FuturesExecutorMixin
 
 
 class MockFuturesExecutor(FuturesExecutorMixin): ...